requirements.txt
decorator==5.0.9
Flask==2.0.1
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
flask_login==0.6.2
flask_bcrypt==1.0.1
pyotp==2.7.0
pyjwt==2.6.0
requests
pandas